Factor out request for label templates
